software/coral/requirements.txt
cobs==1.2.0
fastapi==0.95.0
numpy==1.16.2
opencv_contrib_python==4.7.0.72
pyserial==3.5
uvicorn==0.21.1
